+++ core/src/main/java/com/sigmundgranaas/forgero/texture/utils/Offset.java
@@ -2,8 +2,8 @@ package com.sigmundgranaas.forgero.texture.utils;
 
 import java.util.List;
 
-public record Offset(int x, int y) {
-    public static Offset of(List<Integer> offSetList) {
+public record Offset(float x, float y) {
+    public static Offset of(List<Float> offSetList) {
         if (offSetList == null || offSetList.size() < 2) {
             return new Offset(0, 0);
         }
